+++ p5-server/src/bin/p5-server.ts
@@ -76,6 +76,7 @@ program
   .option('-o, --open', 'Open the page in a browser')
   .option('-p, --port [PORT]', 'HTTP port to listen on', '3000')
   .option('-t, --theme [FILE]', 'template file')
+  .option('--split', 'Use the split (directory + sketch) template')
   .option(
     '--console [FORMAT]',
     'Relay console messages and errors to sketch in the server console'